+++ recipes/install-custom-metrics.rb
@@ -5,6 +5,7 @@
 include_recipe "mh-opsworks-recipes::create-metrics-dependencies"
 
 aws_instance_id = node[:opsworks][:instance][:aws_instance_id]
+stack_id = node[:opsworks][:stack][:id]
 
 cookbook_file "disk_free_metric.sh" do
   path "/usr/local/bin/disk_free_metric.sh"
@@ -41,12 +42,29 @@ cookbook_file "memory_used_metric.sh" do
   mode "755"
 end
 
+cookbook_file "instances_started_ok_metric.sh" do
+  path "/usr/local/bin/instances_started_ok_metric.sh"
+  owner "root"
+  group "root"
+  mode "755"
+end
+
 cron_d 'disk_metrics' do
   user 'custom_metrics'
   minute '*/2'
   # Redirect stderr and stdout to logger. The command is silent on succesful runs
   command %Q(/usr/local/bin/disk_free_metric.sh "#{aws_instance_id}" "type ext|type xfs" 2>&1 | logger -t info)
   path '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin'
+end
+
+if monitoring_node?
+  cron_d 'instances_started_ok_metrics' do
+    user 'custom_metrics'
+    minute '*/2'
+    # Redirect stderr and stdout to logger. The command is silent on succesful runs
+    command %Q(/usr/local/bin/instances_started_ok_metric.sh "#{stack_id}" 2>&1 | logger -t info)
+    path '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin'
+  end
 end
 
 if admin_node?